Task runner default tasks now essentially alias the build task. The original default task of watching for changes has now been moved out to a dedicated watch task.
