+++ run/tasks/default/gulp.js
@@ -1,51 +1,15 @@
 'use strict';
 
 /**
- *  Watches specific types of assets, or all, based on a CLI parameter.
- *  When any of these source files associated with an asset type changes
- *  this triggers compilation methods.
+ *  The default method of the task runner triggers a build task.
  *
  *  Example Usage:
  *  gulp
- *  gulp --watchType styles
- *  gulp --watchType styles,templates
  */
 
 var gulp = require('gulp'),
-    args = require('yargs').argv,
-    common = require('./_common'),
-    globalSettings = require('../../_global');
+    args = require('yargs').argv;
 
 gulp.task('default', function() {
-    var watchFunctions = {
-        styles: function() {
-            console.log('Watching styles...');
-            gulp.watch(common.watchPaths.styles, ['styles']);
-        },
-        scripts: function() {
-            console.log('Watching scripts...');
-            gulp.watch(common.watchPaths.scripts, ['scripts']);
-        },
-        templates: function() {
-            console.log('Watching templates...');
-
-            // RequireJS templates need to be compiled first, before triggering scripts.
-            if (globalSettings.moduleFormat === 'requirejs') {
-                gulp.watch(common.watchPaths.templates, ['templateWatcher']);
-            } else {
-                gulp.watch(common.watchPaths.templates, ['scripts']);
-            }
-        }
-    };
-
-    common.setupWatchers(args, watchFunctions);
-});
-
-/**
- *  This task exists to ensure that templates are called and completed
- *  before attempting to build JS files. Gulp would run these two methods
- *  in tandem and the build file wouldn't pick up the new template files.
- */
-gulp.task('templateWatcher', ['templates'], function() {
-    gulp.start('scripts');
+    gulp.start('build');
 });
