+++ run/tasks/watch/grunt.js
@@ -1,0 +1,62 @@
+'use strict';
+
+/**
+ *  Watches specific types of assets, or all, based on a CLI parameter.
+ *  When any of these source files associated with an asset type changes
+ *  this triggers compilation methods.
+ *
+ *  Example Usage:
+ *  grunt watch
+ *  grunt watch --watchType styles
+ *  grunt watch --watchType styles,templates
+ */
+
+var grunt = require('grunt'),
+    args = require('yargs').argv,
+    common = require('./_common'),
+    globalSettings = require('../../_global');
+
+grunt.loadNpmTasks('grunt-contrib-watch');
+
+grunt.config.set('watch', {
+    options: {
+        interrupt: true
+    }
+});
+
+grunt.registerTask('watch', function() {
+    var watchFunctions = {
+        styles: function() {
+            console.log('Watching styles...');
+            grunt.config.set('watch.styles', {
+                files: common.watchPaths.styles,
+                tasks: ['styles']
+            });
+        },
+        scripts: function() {
+            console.log('Watching scripts...');
+            grunt.config.set('watch.scripts', {
+                files: common.watchPaths.scripts,
+                tasks: ['scripts']
+            });
+        },
+        templates: function() {
+            console.log('Watching templates...');
+
+            var specificTasks = null;
+            if (globalSettings.moduleFormat === 'requirejs') {
+                specificTasks = ['templates', 'scripts'];
+            } else {
+                specificTasks = ['scripts'];
+            }
+
+            grunt.config.set('watch.templates', {
+                files: common.watchPaths.templates,
+                tasks: specificTasks
+            });
+        }
+    };
+
+    common.setupWatchers(args, watchFunctions);
+    grunt.task.run(['watch']);
+});
